+++ src/main/java/application/JobAssignmentAPI/Repositories/HumanResourceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import application.JobAssignmentAPI.Entities.HumanResourceEntity;
 
-public interface HumanResourceRepository extends JpaRepository<HumanResourceEntity, Long>{
+public interface HumanResourceRepository extends JpaRepository<HumanResourceEntity, Integer>{
 
 }
