+++ src/main/java/application/JobAssignmentAPI/Repositories/JobRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import application.JobAssignmentAPI.Entities.JobEntity;
 
-public interface JobRepository extends JpaRepository<JobEntity, Long> {
+public interface JobRepository extends JpaRepository<JobEntity, Integer> {
 
 }
